requirements.txt
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
flake8==3.8.3
Flask==1.1.2
httpie==2.1.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
